+++ examples/llms/litellm_complete.py
